Добавляет интерактивность на страницу избранного, группирует и отображает оферы по городам. Использует один компонент карточки объявления для отрисовки на главной и на странице избранного.
