+++ src/pages/favorites/favorites.tsx
@@ -2,8 +2,7 @@ import Logo from 'components/logo/logo.tsx';
 import Header from 'components/header/header.tsx';
 import Layout from 'components/layout/layout.tsx';
 import {OfferFromList} from 'types/offer.ts';
-import FavoritesCardList from 'components/favorites-card-list/favorites-card-list.tsx';
-import FavoritesList from '../../components/favorites-list/favorites-list.tsx';
+import FavoritesList from 'components/favorites-list/favorites-list.tsx';
 
 type FavoritesProps={
   offers: OfferFromList[];
@@ -17,18 +16,6 @@ function Favorites({offers}: FavoritesProps): JSX.Element {
           <div className="page__favorites-container container">
             <section className="favorites">
               <h1 className="favorites__title">Saved listing</h1>
-              {/*<ul className="favorites__list">*/}
-              {/*  /!*<li className="favorites__locations-items">*!/*/}
-              {/*  /!*  <div className="favorites__locations locations locations--current">*!/*/}
-              {/*  /!*    <div className="locations__item">*!/*/}
-              {/*  /!*      <a className="locations__item-link" href="#">*!/*/}
-              {/*  /!*        <span>Paris</span>*!/*/}
-              {/*  /!*      </a>*!/*/}
-              {/*  /!*    </div>*!/*/}
-              {/*  /!*  </div>*!/*/}
-              {/*  /!*  <FavoritesCardList offers={offers}/>*!/*/}
-              {/*  /!*</li>*!/*/}
-              {/*</ul>*/}
               <FavoritesList offers={offers} />
             </section>
           </div>
